Added room rate validations (Task 2 fixes)
