+++ lib/main.dart
@@ -6,9 +6,8 @@ import 'package:Zonova_Mist/src/core/i18n/arb/app_localizations.dart';
 import 'package:Zonova_Mist/src/core/routing/app_router.dart';
 import 'package:Zonova_Mist/src/core/theme/app_theme.dart';
 import 'package:Zonova_Mist/src/features/auth/login_screen.dart';
-import 'package:Zonova_Mist/src/features/home/rooms/home_screen.dart';
+import 'package:Zonova_Mist/src/features/home/home_screen.dart';
 import 'package:Zonova_Mist/src/shared/widgets/splash_screen.dart';
-import 'package:Zonova_Mist/src/features/home/rooms/room_rate_page.dart';
 
 void main() {
   runApp(
@@ -33,13 +32,8 @@ class MyApp extends ConsumerWidget {
       navigatorKey: AppRouter.navigatorKey,
       debugShowCheckedModeBanner: false,
 
-      // 🔥 FIXED: Dashboard route එක add කරලා
-      routes: {
-        '/dashboard': (context) => const HomeScreen(), // 🎯 මේක තමයි dashboard
-        '/room-rate': (context) => const RoomRatePage(),
-      },
 
-      // Auth state අනුව home screen එක select කරනවා
+
       home: authState.when(
         loading: () => const SplashScreen(),
         authenticated: (_) => const HomeScreen(),
