+++ lib/src/features/home/rooms/room_details_screen.dart
@@ -4,6 +4,7 @@ import 'package:intl/intl.dart';
 import '../../../core/api/api_service.dart';
 import '../../../shared/widgets/common_image_manager.dart';
 import 'edit_room_screen.dart';
+import 'room_rate_page.dart';
 
 class RoomDetailsScreen extends ConsumerStatefulWidget {
   final Map<String, dynamic> room;
@@ -19,7 +20,6 @@ class _RoomDetailsScreenState extends ConsumerState<RoomDetailsScreen> {
   Widget build(BuildContext context) {
     final roomId = widget.room['_id'] as String? ?? '';
 
-    // Debug pricePerNight
     print('Room Price Per Night: ${widget.room['pricePerNight']}');
 
     return Scaffold(
@@ -58,7 +58,6 @@ class _RoomDetailsScreenState extends ConsumerState<RoomDetailsScreen> {
                 child: Column(
                   crossAxisAlignment: CrossAxisAlignment.start,
                   children: [
-                    // Room info
                     Row(
                       children: [
                         Icon(Icons.meeting_room, color: Colors.blue.shade700),
@@ -83,8 +82,7 @@ class _RoomDetailsScreenState extends ConsumerState<RoomDetailsScreen> {
                       children: [
                         const Icon(Icons.group, size: 18, color: Colors.grey),
                         const SizedBox(width: 6),
-                        Text(
-                            'Max Occupancy: ${widget.room['maxOccupancy'] ?? 'N/A'}'),
+                        Text('Max Occupancy: ${widget.room['maxOccupancy'] ?? 'N/A'}'),
                       ],
                     ),
                     const SizedBox(height: 8),
@@ -92,19 +90,26 @@ class _RoomDetailsScreenState extends ConsumerState<RoomDetailsScreen> {
                       children: [
                         const Icon(Icons.attach_money, size: 18, color: Colors.grey),
                         const SizedBox(width: 6),
-                        Text(
-                            'Price Per Night: LKR ${widget.room['pricePerNight'] ?? 'N/A'}'),
+                        Text('Price Per Night: LKR ${widget.room['pricePerNight'] ?? 'N/A'}'),
                       ],
                     ),
 
-                    // Room Rate Display View
-                    RoomRateDisplay(
-                      basePrice: widget.room['pricePerNight'] != null
-                          ? (widget.room['pricePerNight'] as num).toDouble()
-                          : 0,
+                    const SizedBox(height: 15),
+
+                    ElevatedButton(
+                      onPressed: () {
+                        Navigator.push(
+                          context,
+                          MaterialPageRoute(
+                            builder: (_) => const RoomRatePage(),
+                          ),
+                        );
+                      },
+                      child: const Text("View Room Rates"),
                     ),
 
-                    const SizedBox(height: 8),
+                    const SizedBox(height: 15),
+
                     Row(
                       children: [
                         const Icon(Icons.check_circle, size: 18, color: Colors.grey),
@@ -123,8 +128,7 @@ class _RoomDetailsScreenState extends ConsumerState<RoomDetailsScreen> {
                     const SizedBox(height: 12),
                     const Text(
                       'Amenities',
-                      style:
-                      TextStyle(fontWeight: FontWeight.bold, fontSize: 16),
+                      style: TextStyle(fontWeight: FontWeight.bold, fontSize: 16),
                     ),
                     const SizedBox(height: 4),
                     Text(
@@ -146,115 +150,3 @@ class _RoomDetailsScreenState extends ConsumerState<RoomDetailsScreen> {
     );
   }
 }
-
-// Room Rate Display Widget
-class RoomRateDisplay extends StatefulWidget {
-  final double basePrice;
-
-  const RoomRateDisplay({super.key, required this.basePrice});
-
-  @override
-  State<RoomRateDisplay> createState() => _RoomRateDisplayState();
-}
-
-class _RoomRateDisplayState extends State<RoomRateDisplay> {
-  DateTime? checkInDate;
-  int nights = 1;
-  int rooms = 1;
-  int people = 1;
-
-  double calculateRate() {
-    double extraPersonRate = 20;
-    return (widget.basePrice * nights * rooms) +
-        ((people - 1) * extraPersonRate * nights * rooms);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      elevation: 2,
-      margin: const EdgeInsets.symmetric(vertical: 10),
-      child: Padding(
-        padding: const EdgeInsets.all(12.0),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            // Check-in date picker
-            Row(
-              children: [
-                const Text('Check-in: '),
-                TextButton(
-                  child: Text(checkInDate != null
-                      ? DateFormat('MMM dd, yyyy').format(checkInDate!)
-                      : 'Select Date'),
-                  onPressed: () async {
-                    DateTime? picked = await showDatePicker(
-                      context: context,
-                      initialDate: DateTime.now(),
-                      firstDate: DateTime.now(),
-                      lastDate: DateTime.now().add(const Duration(days: 365)),
-                    );
-                    if (picked != null) {
-                      setState(() {
-                        checkInDate = picked;
-                      });
-                    }
-                  },
-                ),
-              ],
-            ),
-
-            // Nights
-            Row(
-              children: [
-                const Text('Nights: '),
-                DropdownButton<int>(
-                  value: nights,
-                  items: List.generate(30, (index) => index + 1)
-                      .map((e) => DropdownMenuItem(value: e, child: Text('$e')))
-                      .toList(),
-                  onChanged: (val) => setState(() => nights = val!),
-                ),
-              ],
-            ),
-
-            // Rooms
-            Row(
-              children: [
-                const Text('Rooms: '),
-                DropdownButton<int>(
-                  value: rooms,
-                  items: List.generate(10, (index) => index + 1)
-                      .map((e) => DropdownMenuItem(value: e, child: Text('$e')))
-                      .toList(),
-                  onChanged: (val) => setState(() => rooms = val!),
-                ),
-              ],
-            ),
-
-            // People
-            Row(
-              children: [
-                const Text('People: '),
-                DropdownButton<int>(
-                  value: people,
-                  items: List.generate(10, (index) => index + 1)
-                      .map((e) => DropdownMenuItem(value: e, child: Text('$e')))
-                      .toList(),
-                  onChanged: (val) => setState(() => people = val!),
-                ),
-              ],
-            ),
-
-            const SizedBox(height: 12),
-            Text(
-              'Total Price: LKR ${calculateRate().toStringAsFixed(2)}',
-              style: const TextStyle(fontWeight: FontWeight.bold, fontSize: 16),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
